requirements.txt
# ── Claude API ──────────────────────────────
anthropic>=0.40.0           # Claude SDK，支持 prompt caching

# ── PDF 解析 ────────────────────────────────
pymupdf>=1.24.0             # PDF 解析（速度快、结构提取好）

# ── EPUB 解析 ───────────────────────────────
ebooklib>=0.18              # EPUB 读取
beautifulsoup4>=4.12.0      # HTML 内容清洗
lxml>=5.0.0                 # BS4 高性能解析器

# ── Token 计数 ──────────────────────────────
tiktoken>=0.7.0             # Token 估算

# ── 数据模型与验证 ──────────────────────────
pydantic>=2.7.0             # 数据模型定义与验证

# ── CLI 与交互 ──────────────────────────────
click>=8.1.0                # CLI 框架
rich>=13.7.0                # 终端富文本、进度条
questionary>=2.0.0          # 交互式命令行问答

# ── 模板渲染 ────────────────────────────────
jinja2>=3.1.0               # Markdown 报告模板引擎

# ── 配置管理 ────────────────────────────────
python-dotenv>=1.0.0        # .env 文件加载
pyyaml>=6.0.1               # config.yaml 读取

# ── 工具库 ──────────────────────────────────
tenacity>=8.3.0             # 重试机制（指数退避）
diskcache>=5.6.0            # 本地磁盘缓存
loguru>=0.7.0               # 结构化日志
python-slugify>=8.0.0       # 文件名安全转换
httpx>=0.27.0               # 异步 HTTP
